requirements.txt
django==2.2.7
django-jet==1.0.8
whitenoise==3.3.1
gunicorn==19.9.0
dj-database-url==0.5.0
psycopg2-binary==2.8.4
django-solo
